[recipe] feat: add retool recipe
