+++ recipe/retool/test_chat_scheduler.py
@@ -1,0 +1,154 @@
+# Copyright 2024 Bytedance Ltd. and/or its affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import asyncio
+import os
+import socket
+import sys
+import tempfile
+from contextlib import asynccontextmanager
+
+import fastapi
+import numpy as np
+import ray
+import uvicorn
+from datasets import load_dataset
+from omegaconf import OmegaConf
+from starlette.requests import Request
+from starlette.responses import JSONResponse
+
+from tests.workers.rollout.async_rollout_utils import init_async_rollout_manager
+from verl.protocol import DataProto
+
+
+def _get_free_port():
+    with socket.socket() as sock:
+        sock.bind(("", 0))
+        return sock.getsockname()[1]
+
+
+@ray.remote(num_cpus=1)
+class Sandbox:
+    """Sandbox to execute python code.
+
+    WARNING: This class is for testing purpose only, do not use it in production.
+    Please use a sandbox with strong isolation and security restrictions instead.
+    """
+
+    def __init__(self):
+        self.address = ray._private.services.get_node_ip_address()
+        self.port = None
+        self.server_ready = asyncio.Event()
+        asyncio.create_task(self._start_fastapi_server())
+
+    async def code_execution(self, request: Request):
+        request_json = await request.json()
+        code = request_json["code"]
+        print(f"execute code:\n{code}")
+
+        _, temp_file = tempfile.mkstemp(suffix=".py", prefix="temp_code", dir=None, text=True)
+        with open(temp_file, "w") as f:
+            f.write(code)
+
+        try:
+            process = await asyncio.create_subprocess_exec(sys.executable, temp_file, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE)
+
+            stdout, stderr = await process.communicate()
+
+            return JSONResponse(content={"stdout": stdout.decode(), "stderr": stderr.decode(), "returncode": process.returncode})
+        finally:
+            try:
+                os.unlink(temp_file)
+            except:  # noqa: E722
+                pass
+
+    async def _start_fastapi_server(self):
+        @asynccontextmanager
+        async def lifespan(app: fastapi.FastAPI):
+            print("FastAPI startup")
+            self.server_ready.set()
+            yield
+
+            print("FastAPI shutdown, maybe address already in use, exit process immediately.")
+            os._exit(-1)
+
+        app = fastapi.FastAPI(lifespan=lifespan)
+        app.router.add_api_route("/run_code", self.code_execution, methods=["POST"])
+
+        self.port = _get_free_port()
+        config = uvicorn.Config(app, host=["::", "0.0.0.0"], port=self.port, log_level="warning")
+        server = uvicorn.Server(config)
+        await server.serve()
+
+    async def get_server_address(self) -> str:
+        """Get FastAPI server address."""
+        await self.server_ready.wait()
+        return f"{self.address}:{self.port}"
+
+
+user_prompt_template = """
+You are a helpful assistant. Let's solve math problem in following steps:
+1. Write a python code first and return the code to user, the code must be in following format:
+
+<code>
+```python
+import os
+
+print(...)
+```
+</code>
+
+The code must explictly print necessary output to stdout. Remember stop generation at </code> immediately and return the code.
+2. User will send the python code to a external sandbox to execute and get output from stdout.
+3. User will send the output in format <interpreter>output</interpreter> to you, and you should use the output to answer the question.
+The answer format must be: <answer>\\boxed{'The final answer goes here.'}</answer>
+
+*user question:*
+{question}
+"""
+
+
+if __name__ == "__main__":
+    ray.init(
+        runtime_env={
+            "env_vars": {
+                "TOKENIZERS_PARALLELISM": "true",
+                "NCCL_DEBUG": "WARN",
+                "VLLM_LOGGING_LEVEL": "INFO",
+                "VLLM_USE_V1": "1",
+            }
+        }
+    )
+
+    # Load config
+    config = OmegaConf.load("verl/trainer/config/ppo_trainer.yaml")
+    config.actor_rollout_ref.model.path = "Qwen/Qwen2.5-1.5B-Instruct"
+    config.actor_rollout_ref.rollout.mode = "async"
+    config.actor_rollout_ref.rollout.chat_scheduler = "recipe.retool.chat_scheduler.ToolChatCompletionScheduler"
+    config.actor_rollout_ref.rollout.prompt_length = 8192
+    config.actor_rollout_ref.rollout.response_length = 8192
+
+    # Init sandbox and async rollout manager
+    sandbox = Sandbox.options(num_cpus=1).remote()
+    sandbox_address = ray.get(sandbox.get_server_address.remote())
+    async_rollout_manager = init_async_rollout_manager(
+        config,
+        scheduler_kwargs={"sandbox_url": f"http://{sandbox_address}/run_code", "user_prompt_template": user_prompt_template},
+    )
+
+    # Build dataset
+    dataset = load_dataset("Maxwell-Jia/AIME_2024", split="train")
+    prompts = DataProto(non_tensor_batch={"raw_prompt": np.array([[{"role": "user", "content": problem}] for problem in dataset["Problem"]])})
+
+    result = async_rollout_manager.generate_sequences(prompts=prompts)
+    assert len(result) == len(dataset)
